changed all odometry to double. it works already a lot better. but still a lot WTF

+++ cvra_cs.c
@@ -42,6 +42,7 @@ struct _rob robot;
 
 /** Limite PWM = + ou - 475 */
 
+
 void cvra_cs_init(void) {
     /****************************************************************************/
     /*                                Motor                                     */
@@ -142,15 +143,15 @@ void cvra_cs_init(void) {
 
     holonomic_position_init(&robot.pos);
 
-    float beta[] = {ROBOT_BETA_WHEEL0_RAD,
+    double beta[] = {ROBOT_BETA_WHEEL0_RAD,
                     ROBOT_BETA_WHEEL1_RAD,
                     ROBOT_BETA_WHEEL2_RAD};
 
-    float wheel_radius[] = {ROBOT_RADIUS_WHEEL0_MM,
+    double wheel_radius[] = {ROBOT_RADIUS_WHEEL0_MM,
                             ROBOT_RADIUS_WHEEL1_MM,
                             ROBOT_RADIUS_WHEEL2_MM};
 
-    float wheel_distance[] = {ROBOT_DISTANCE_WHEEL0_MM,
+    double wheel_distance[] = {ROBOT_DISTANCE_WHEEL0_MM,
                               ROBOT_DISTANCE_WHEEL1_MM,
                               ROBOT_DISTANCE_WHEEL2_MM};
 
@@ -172,6 +173,7 @@ void cvra_cs_init(void) {
                                      HEXMOTORCONTROLLER_BASE};
 
     holonomic_position_set_mot_encoder(&robot.pos, motor_encoder, motor_encoder_param);
+
 
     /****************************************************************************/
     /**      CS pour les macros-variables (seulement les rampes, pas de PID)    */
